requirements.txt
numpy==1.24.1
Pillow==9.4.0
pydicom==2.3.1
pynetdicom==2.0.2
